+++ src/Rtcp/RtcpFCI.h
@@ -353,7 +353,7 @@ public:
     //单位64ms
     uint32_t getReferenceTime() const;
     uint16_t getPacketCount() const;
-    map<uint16_t, std::pair<SymbolStatus, uint32_t/*recv delta,单位为250us*/> > getPacketChunkList(size_t total_size) const;
+    map<uint16_t/*rtp ext seq*/, std::pair<SymbolStatus, int16_t/*recv delta,单位为250us*/> > getPacketChunkList(size_t total_size) const;
 
 private:
     //base sequence number,基础序号,本次反馈的第一个包的序号;也就是RTP扩展头的序列号
